feat: api 연결 및 data type 오류 해결

+++ app/src/main/java/com/example/ourmenu/data/map/data/MapData.kt
@@ -2,8 +2,8 @@ package com.example.ourmenu.data.map.data
 
 data class MapData(
     val groupId: Int,
-    val latitude: Int,
-    val longitude: Int,
+    val latitude: Double,
+    val longitude: Double,
     val menuIconType: String,
     val menuTitle: String,
     val placeId: Int,
@@ -18,8 +18,8 @@ data class MapSearchData(
 
 data class MapInfoDetailData(
     val groupId: Int,
-    val latitude: Int,
-    val longitude: Int,
+    val latitude: Double,
+    val longitude: Double,
     val menuFolder: MenuFolder,
     val menuIconType: String,
     val menuImgsUrl: ArrayList<MenuImgsUrl>,
